### 2023-10-02 #5918 @frank1010111
Hello! A few potential reviewers include Dennis Glaeser (dglaeser), Guilherme P. Castelao (castelao), wkearn, leouieda and sgrieve.

Is 1200 lines of code on the low side?

I used Tokei to calculate the lines of code, which appears to handle docstrings as code, while your tool calls them comments:
```
tokei
===============================================================================
 Language            Files        Lines         Code     Comments       Blanks
===============================================================================
 Dockerfile              1           16           12            1            3
 Python                 11         1831         1614           27          190
 Rust                    2          164          150            2           12
 TOML                    2          132          114            3           15
 Jupyter Notebooks       4            0            0            0            0
 |- Markdown             3           71            1           49           21
 |- Python               4          321          290            3           28
 (Total)                            392          291           52           49
 Markdown               11          521            0          361          160
 |- BASH                 1            1            1            0            0
 |- Python               1           12           12            0            0
 (Total)                            534           13          361          160
===============================================================================
 Total                  31         2664         1890          394          380
===============================================================================
```
