### 2023-10-02 #5918 @kthyng
Yes 1000 is the approximate cut off for needing to query scope: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort
