+++ pyproject.toml
@@ -47,10 +47,12 @@ build-backend = "maturin"
 
 [tool.maturin]
 # Include arbitrary files in the sdist
-sdist-include = []
+# sdist-include = []
 compatibility = "manylinux2014"
 # Bindings type
 bindings = "pyo3"
+python-source = "python"
+module-name = "pywaterflood._core"
 
 [tool.coverage.run]
 branch = true
