+++ .github/pull_request_template.md
@@ -10,7 +10,7 @@ It enhances ...
 
 # Pull request checklist
 
-- [ ] If features have changed, there's new documentation, and it has been checked: `nox -s docs -- serve`
+- [ ] If features have changed, there's new documentation, and it has been checked: `nox -s docs -- --serve`
 - [ ] If a bugfix, new tests are in `testing/`
 - [ ] Passes all CI/CD tests:
   - [ ] Pre-commit linting passes: `nox -s lint`
