### 2023-10-06 235d20c
:memo: :bug: :hammer: change from myst_nb to myst_parser for docs to fix a deprecation in sphinx-autoapi
