+++ pyproject.toml
@@ -27,8 +27,8 @@ dependencies = [
 test = ["pytest >=6.2.5", "pytest-cov >=3.0.0", "coverage[toml] >=6.3.3"]
 docs = [
     "furo",
-    "myst-nb",
-    "sphinx",
+    "myst_parser>=0.13",
+    "sphinx>=7.0",
     "sphinx-autoapi",
     "sphinx-copybutton",
     "sphinx-design",
