### 2023-10-08 941ecbc
:memo: Update CHANGELOG for v0.3.2
