### 2023-10-08 4403f2c
:ambulance: switch sphinx extension to myst-parser
