### 2023-10-09 5bf613b
:bug: :memo: nbsphinx+pandoc to build locally
