+++ docs/conf.py
@@ -19,6 +19,7 @@ author = "Frank Male"
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
+    "nbsphinx",
     "myst_parser",
     "autoapi.extension",
     "sphinx.ext.mathjax",
