### 2023-10-09 8cd7418
:memo: add pandoc dependency for readthedocs
