+++ .readthedocs.yml
@@ -7,8 +7,10 @@ sphinx:
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.9"
-    rust: "1.64"
+    python: "3.11"
+    rust: "1.73"
+  apt_packages:
+    - pandoc
 
 python:
   install:
