### 2023-10-10 a5766d7
:memo: update rust version for readthedocs
