+++ .readthedocs.yml
@@ -8,7 +8,7 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.11"
-    rust: "1.73"
+    rust: "1.70"
   apt_packages:
     - pandoc
 
