### 2023-10-10 b8396c0
:memo: add Buckley-Leverett example notebook to user guide
