+++ docs/user-guide/index.md
@@ -9,4 +9,5 @@ New users should start with {doc}`7-minutes-to-pywaterflood`.
 7-minutes-to-pywaterflood.ipynb
 choosing-crm.ipynb
 multiwell-productivity-index.ipynb
+buckley-leverett.ipynb
 ```
