### 2023-10-12 #5918 @frank1010111
Okay, how does a scope query work?
