### 2023-10-25 c22059f
:sparkles: marsal and klins approximations for dimensionless water influx
