### 2023-11-08 d1f33dc
:hook: GHA: check wheel contents before publishing to TestPyPI or PyPI
