### 2023-11-23 #5918 @elbeejay
@frank1010111 I took a look and the code and paper look pretty good to me, so I'll start looking for reviewers. I'll likely wait until next week to start pinging folks due to the holiday here in the US.

My one note on the paper is that you should introduce the acronym "CRM" as a parenthetical "Capacitance Resistance Models (CRM)" (maybe on line 7) as later the acronym is used (line 15).
