### 2023-11-28 #5918 @elbeejay
@davidlitwin would you be interested in reviewing "Pywaterflood"? I know you are familiar with the JOSS review process so please take a look when you get a chance and let us know if you think you'll have the time to conduct a review.

Hope you are well,
Jay
