### 2023-12-05 #5918 @elbeejay
@mgcooper might you be interested in conducting a review for JOSS? I'm not sure how comfortable you are with Python, but the subject area might be a good fit, the submission is titled: "Pywaterflood: Well connectivity analysis through capacitance-resistance modeling".

Please let me know if you'd be willing to review, thanks!
