### 2023-12-05 #5918 @elbeejay
@katmratliff would you be interested in conducting a review for JOSS? The submission is titled: "Pywaterflood: Well connectivity analysis through capacitance-resistance modeling".  It might be a bit outside of your typical sub-discipline, but I think you might find it interesting and tangentially related to some of your work. I know you've published with and reviewed for JOSS in the past so will skip the spiel about the review process - please just let me know if you'd like to review, thanks!
