### 2023-12-14 #5918 @elbeejay
@mancellin, any chance you'd be interested in reviewing this submission to JOSS titled: "Pywaterflood: Well connectivity analysis through capacitance-resistance modeling"? I think this submission falls (broadly) within your area of expertise. I know you've published with and reviewed for JOSS in the past so I thought I'd reach out. Let me know if you'd be available and interested in reviewing -- no pressure at all if not!

Thanks,
Jay
