### 2023-12-20 #5918 @mancellin
> @mancellin, any chance you'd be interested in reviewing this submission to JOSS titled: "Pywaterflood: Well connectivity analysis through capacitance-resistance modeling"? I think this submission falls (broadly) within your area of expertise. I know you've published with and reviewed for JOSS in the past so I thought I'd reach out. Let me know if you'd be available and interested in reviewing -- no pressure at all if not!
>
> Thanks, Jay

Thank you for the invitation. The package seems to be very domain-specific and not an application I'm very familiar with. But I can give it a try. You might need to be patient as I won't be able to start working on this before mid-January due to holidays.
