### 2024-01-01 94cd86e
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 23.7.0 → 23.12.1](https://github.com/psf/black/compare/23.7.0...23.12.1)
- [github.com/pre-commit/pre-commit-hooks: v4.4.0 → v4.5.0](https://github.com/pre-commit/pre-commit-hooks/compare/v4.4.0...v4.5.0)
- https://github.com/charliermarsh/ruff-pre-commit → https://github.com/astral-sh/ruff-pre-commit
- [github.com/astral-sh/ruff-pre-commit: v0.0.284 → v0.1.9](https://github.com/astral-sh/ruff-pre-commit/compare/v0.0.284...v0.1.9)
- [github.com/pre-commit/mirrors-prettier: v3.0.1 → v4.0.0-alpha.8](https://github.com/pre-co
[…]
