### 2024-01-03 #5918 @elbeejay
> > @mgcooper might you be interested in conducting a review for JOSS? I'm not sure how comfortable you are with Python, but the subject area might be a good fit, the submission is titled: "Pywaterflood: Well connectivity analysis through capacitance-resistance modeling".
> > Please let me know if you'd be willing to review, thanks!
>
> @elbeejay In general I am happy to review for JOSS and I am comfy with python. In this particular case, I could complete the review by ~8 February. I will be unavailable from 9 Feb to 3 March. Please reach out or feel free to assign me if this timeline is acceptable.

That'd be great and very much appreciated, thanks! The timeline works just fine.
