### 2024-01-06 #5918 @elbeejay
@dankusanovic would you be interested in reviewing (once again) for JOSS. The submission is titled: "Pywaterflood: Well connectivity analysis through capacitance-resistance modeling" and the submitting author, @frank1010111, believes you'd be able to provide a good review.

Please let us know, thank you!
