### 2024-01-07 #5918 @amandersillinois
Yes - I can provide this review – I've been on vacation and missed your previous emails.

-Alison

ALISON M ANDERS
Associate Professor

Department of Geology
College of Liberal Arts and Sciences Administration
University of Illinois at Urbana-Champaign
Dept. of Geology
3081 Natural History Building | M/C 102
Urbana, IL 61801
217.244.3917 | ***@***.******@***.***>
www.geology.illinois.edu<http://www.geology.illinois.edu/>

she/her pronouns

[cid:74ecaa58-3c61-4cbc-bd8c-8ffd0d44235c]<http://illinois.edu/>

Under the Illinois Freedom of Information Act any written communication to or from university employees regarding university business is a public record and may be subject to public disclosure.
