### 2024-01-08 #5918 @elbeejay
@mancellin, given your availability but more importantly given that you are not familiar with the application and specific domain being presented here, I've not added you to the list of reviewers for this submission. I appreciate your willingness to review for JOSS and will be sure to reach out when I see a submission that is a better fit for your expertise.
