### 2024-01-08 #6191 @editorialbot
[REVIEW]: Pywaterflood: Well connectivity analysis through capacitance-resistance modeling
**Submitting author:** @frank1010111 (<a href="http://orcid.org/0000-0002-3402-5578">Frank Male</a>)
**Repository:** https://github.com/frank1010111/pywaterflood
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.3
