### 2024-01-08 #6191 @elbeejay
Hi @mgcooper & @amandersillinois, thank you both again for agreeing to review for JOSS. This is the review issue where we track the peer-review process. Please refer to the first post in this issue, as it does a pretty good job outlining the JOSS review procedure; we do "checklist-driven" reviews, and there is a command in the first post for you to use to generate your own reviewer checklist.

At this time we are asking reviewers to please complete their first set of reviews within 6 weeks; this is flexible, really what I and I'm sure @frank1010111 would appreciate is if you just comment here with roughly when you expect to get to your review and/or if you anticipate a delay in reviewing this submission. @mgcooper I know you'd indicated you are busy after Feb. 8 which is great for us to know.

Please do not hesistate to reach out to me with any questions you might have about the review process. I will be setting up reminders for the bot to ping us all in 3 weeks just to help make sure this review doesn't drop off of anyone's to-do list.

Thanks all,
Jay
