### 2024-01-16 c43d0f2
:bug: :hook: use rust pre-commit hook from working pull request
solves issue with pre-commit.ci not knowing to install rust
