### 2024-01-20 #6191 @elbeejay
Hi all, just wanted to check-in and ask @mgcooper and @amandersillinois if y'all have any questions about the JOSS review process. Please reach out and ask if so. Thanks!
