+++ .github/workflows/publish-to-pypi.yml
@@ -29,7 +29,7 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Build wheels
-        uses: pypa/cibuildwheel@v2.16.2
+        uses: pypa/cibuildwheel@v2.16.4
 
       - uses: actions/upload-artifact@v3
         with:
