### 2024-02-04 d0382fe
:test_tube: test predicting with CRM without any prior fitting
