### 2024-02-04 ac228a8
:bug: fix predicting without prior fitting
