### 2024-02-04 fd69d14
:memo: added examples and other improvements to docstrings
