+++ python/pywaterflood/crm.py
@@ -1,16 +1,17 @@
 """Analyze waterfloods with capacitance-resistance models.
 
-This is the central module in :code:`pywaterflood`, based around the:code:`CRM`
+This is the central module in ``pywaterflood``, based around the:code:`CRM`
 class, which implements the standard capacitance-resistance models. For most
 cases, the best performance comes from selecting
 :code:`CRM(primary=True, tau_selection="per-pair", constraints="up-to one")`.
-If the data is too sparse, then change :code:`tau_selection` to "per-producer".
+If the data is too sparse, then change ``tau_selection`` to "per-producer".
 
 The base class assumes constant bottomhole pressures for the producing wells.
 If you know the pressures for these wells or at least the trend, consider using
-:code:`CrmCompensated`.
+``CrmCompensated``.
 
 """
+
 from __future__ import annotations
 
 import pickle
@@ -29,7 +30,7 @@ from pywaterflood import _core
 def q_primary(
     production: NDArray, time: NDArray, gain_producer: float, tau_producer: float
 ) -> NDArray:
-    """Calculate primary production contribution.
+    r"""Calculate primary production contribution.
 
     Uses Arps equation with :math:`b=0`
 
@@ -43,14 +44,14 @@ def q_primary(
     time : NDArray
         Producing times to forecast, size: Number of time steps
     gain_producer : float
-        Arps q_i factor
+        Arps :math:`q_i` factor
     tau_producer : float
         Arps time constant
 
     Returns
     ----------
     q_hat : NDArray
-        Calculated production, size: Number of time steps
+        Calculated production, :math:`\hat q`, size: Number of time steps
     """
     return _core.q_primary(production, time, gain_producer, tau_producer)
 
@@ -77,7 +78,7 @@ def q_CRM_perpair(injection: NDArray, time: NDArray, gains: NDArray, taus: NDArr
     Returns
     ----------
     q_hat : NDArray
-        Calculated production, size: Number of time steps
+        Calculated production :math:`\\hat q`, size: Number of time steps
     """
     return _core.q_crm_perpair(injection, time, gains, taus)
 
@@ -85,7 +86,7 @@ def q_CRM_perpair(injection: NDArray, time: NDArray, gains: NDArray, taus: NDArr
 def q_CRM_perproducer(injection: NDArray, time: NDArray, gain: NDArray, tau: float) -> NDArray:
     """Calculate per injector-producer pair production (simplified tank).
 
-    Uses simplified CRMp model that assumes a single tau for each producer
+    Uses simplified CRMP model that assumes a single tau for each producer
 
     Args
     ----------
@@ -102,7 +103,9 @@ def q_CRM_perproducer(injection: NDArray, time: NDArray, gain: NDArray, tau: flo
     Returns
     ----------
     q_hat : NDArray
-        Calculated production, size: Number of time steps
+        Calculated production :math:`\\hat q`
+
+        shape: Number of time steps
     """
     tau2 = np.full(injection.shape[1], tau)
     return q_CRM_perpair(injection, time, gain, tau2)
@@ -128,8 +131,7 @@ def q_bhp(pressure_local: NDArray, pressure: NDArray, v_matrix: NDArray) -> NDAr
     Returns
     -------
     q : NDArray
-        production from changing BHP
-        shape: n_time
+        production from changing BHP, shape: n_time
     """
     return _core.q_bhp(pressure_local, pressure, v_matrix)
 
@@ -139,14 +141,17 @@ def random_weights(n_prod: int, n_inj: int, axis: int = 0, seed: int | None = No
 
     Args
     ----
-    n_i : int
-    n_j : int
+    n_prod : int
+        Number of producing wells
+    n_inj : int
+        Number of injecting wells
     axis : int, default is 0
[…]
