### 2024-02-04 53e0811
:memo: fix example in CRM.predict docstring
