### 2024-02-05 0a13cc1
:art: cleaner prediction functions
