### 2024-02-05 559ecf5
:bug: :white_check_mark: fix prediction tests when no primary production
