+++ docs/user-guide/index.md
@@ -10,4 +10,5 @@ New users should start with {doc}`7-minutes-to-pywaterflood`.
 choosing-crm.ipynb
 multiwell-productivity-index.ipynb
 buckley-leverett.ipynb
+using-docker.md
 ```
