+++ .github/workflows/tests.yml
@@ -39,7 +39,7 @@ jobs:
       - name: Test with nox
         run: pipx run nox -s tests
       - name: "Upload coverage to Codecov"
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v4
         with:
           fail_ci_if_error: true
           token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
