### 2024-02-06 frank1010111/pywaterflood#78 @amandersillinois
Unable to run buckley-leverett.ipynb :bug: 
I have successfully installed pywaterflood and can run 7-minutes-to-pywaterflood.ipynb, choosing-crm.ipynb, and multiwell-productivity-index.ipynb.  However, I can't run /buckley-leverett.ipynb

This is the error I get when trying to run the first cell in the notebook.
Cell In[3], line 6
      3 import seaborn as sns
      4 import pandas as pd
----> 6 from pywaterflood.buckleyleverett import Reservoir, water_front_velocity, breakthrough_sw

ModuleNotFoundError: No module named 'pywaterflood.buckleyleverett'
