### 2024-02-06 #6191 @amandersillinois
I have successfully installed pywaterflood and can run 7-minutes-to-pywaterflood.ipynb, choosing-crm.ipynb, and multiwell-productivity-index.ipynb. However, I can't run /buckley-leverett.ipynb.  I opened an issue on pywaterflood

[https://github.com/frank1010111/pywaterflood/issues/78](url)
