### 2024-02-06 frank1010111/pywaterflood#78 @frank1010111
Thank you for raising this issue. I think the problem is that the `buckleyleverett` submodule has not been released on pywaterflood's PyPI page. Therefore, wheels that have it included have to be hand-rolled. I am nearing the 0.4 release where it will be added.

In the meantime, you can install pywaterflood with the buckleyleverett submodule by following the documentation in the [Contributing guide](https://github.com/frank1010111/pywaterflood/blob/master/CONTRIBUTING.md#get-started).

Or, if you have rust installed on your system, you might be able to run
```
pip install git+https://github.com/frank1010111/pywaterflood.git
```

Let me know if that fixes things, and I hope Buckley-Leverett is useful for you.
