### 2024-02-07 #6191 @amandersillinois
I've done my best to try to follow the guidelines for review but I do have a few questions and some uncertainty about some of the checklist items.

I'm neither a specialist in well connectivity/groundwater nor a particularly adept programmer -- I'm just a user of numerical models. So I don't believe I'm qualified/able to comment two some things.
Specifically: on Functionality-Performance - the article claims that the performance is better than an Excel workbook - but I did not test this claim and don't think I could do so in the time I can devote to this review.
on Documentation -Functionality documentation - I don't know what API method documentation is and cannot comment.

Broadly, on Documentation -- I think that it checks the boxes (other than the question above), but I was sort-of expecting a walk-through of the different options outside of the annotation of the code itself. For example, CRM takes arguments of tau_selection and constraints and the options (and rational for picking one or the other) are not described in detail - although there is advice about what to pick at the top of crm.py.  There is no "user manual" per say - although I appreciate the python notebook examples and I do believe that if I wanted to use this for a specific problem, I would be able to at least get started using what is provided.  One of the notebooks doesn't work (for the current release) without a separate install -but the developer/author helped me very quickly to get it working, so that doesn't appear to be a big issue.

On Quality of writing/Summary - it's generally reasonable, but I think a bit too terse.  As a non-specialist, I have some idea of what problems are being solved, but I think just a few more sentences could make that much more clear. I am hung up on the end of the first sentence -- "geothermal" is an adjective and so requires a noun to describe.  A quick edit from the perspective of improving readability for the non-specialist would, in my opinion, improve the document.
