### 2024-02-07 #6191 @frank1010111
Thank you for your comments, @amandersillinois. To address your points.

1. the API documentation is this part: [latest (unstable) documentation](https://pywaterflood.readthedocs.io/en/latest/autoapi/pywaterflood/index.html) or [stable documentation](https://pywaterflood.readthedocs.io/en/stable/autoapi/pywaterflood/index.html)
2. I take this as an action item to add a walk through of tau_selection, constraints, and options in the [choosing your crm](https://pywaterflood.readthedocs.io/en/stable/user-guide/choosing-crm.html) notebook.
3. I will perform a quick edit of the summary, paying special attention to the problems solved, ease of understanding for a non-specialist, and the word "geothermal".

Does that sound about right?
