### 2024-02-08 5477daf
:memo: edit paper following JOSS review comments
