+++ paper/paper.md
@@ -5,7 +5,7 @@ tags:
   - well connectivity analysis
   - waterfloods
   - CO2 floods
-  - Geothermal
+  - Geothermal energy
   - multiphase flow
 authors:
   - name: Frank Male
@@ -21,17 +21,22 @@ bibliography: paper.bib
 
 # Summary
 
-Well connectivity analysis has many applications for subsurface energy, from waterfloods to CO$_2$ floods to geothermal. Capacitance Resistance Models are useful for performing well connectivity analysis with limited information about the geology of the reservoirs involved. They are so-called because the equations describing well influence mimic a network of capacitors and resistors.
+Well connectivity analysis has many applications for subsurface energy, from waterfloods to CO$_2$ floods to geothermal energy projects. After completing a well connectivity analysis, reservoir managers can better allocate injections and plan well interventions. Capacitance Resistance Models are useful for performing well connectivity analysis with limited information about the geology of the reservoirs involved. They are so-called because the equations describing well influence mimic a network of capacitors and resistors.
 
-`Pywaterflood` is a Python package that uses Capacitance Resistance Modeling to estimate well connectivity. The `CRM` submodule forms the bulk of this package. It can perform capacitance resistance modeling with differing levels of complexity, from assuming that producing and injecting wells share one universal time constant, to each producer has the same time constant with all injectors, to each producer-injector pair has an its own time constant. CRM was developed by @yousef2006capacitance. The `MPI` submodule uses a geometrical model of well influence [@valko2000development], extended and applied to reservoirs with both injecting and producing wells [@kaviani2010inferring].
+`Pywaterflood` is a Python package that uses Capacitance Resistance Modeling (CRM) to estimate well connectivity. The `CRM` submodule forms the bulk of this package. It can perform CRM with differing levels of complexity, from assuming that producing and injecting wells share one universal time constant, to each producer has the same time constant with all injectors, to each producer-injector pair has an its own time constant. CRM was developed by @yousef2006capacitance. The `MPI` submodule uses a geometrical model of well influence [@valko2000development], extended and applied to reservoirs with both injecting and producing wells [@kaviani2010inferring].
 
 # Statement of need
 
-Interwell connectivity analysis is important for understanding the geology of subsurface systems. This can be used to improve oil recovery efficiency [@albertoni2003inferring], better sequester CO$_2$ [@tao2015optimizing], and optimize geothermal fields [@akin2014optimization].
+Interwell connectivity analysis is important for understanding the geology of subsurface systems. This can be used to improve oil recovery efficiency [@albertoni2003inferring], better sequester CO$_2$ [@tao2015optimizing], and optimize geothermal fields [@akin2014optimization]. @holanda2018 enumerate four uses for CRM results:
+
+1. Finding sealing faults and high-flow-connectivity pathways
+2. Investigating connectivity between adjacent reservoirs and reservoir compartments
+3. Measuring the per-well effectiveness of fluid injection
+4. Optimizing injection, either through redirecting fluid to different wells or to inform the placement of new wells
 
 `Pywaterflood` uses a reduced-physics model to match connections between injecting and producing wells. As explained in @holanda2018, capacitance-resistance modeling provides a method for connectivity analysis more sophisticated than empirical decline analysis, but more approachable than full reservoir simulation.
 
-There is another publicly available tool for capacitance resistance modeling reservoirs like `pywaterflood` [@sayarpour2008development]. However, that tool comes in the form of an Excel workbook and no associated license. This python package provides more extensibility and better performance than an Excel file. There are other programs for performing waterflood analysis with CRM in the industry, but they are not open sourced and available for researchers to use.
[…]
