### 2024-02-08 3eb7250
:memo: added explanation in crm.py docstrings
