### 2024-02-09 #6191 @elbeejay
> I'm neither a specialist in well connectivity/groundwater nor a particularly adept programmer -- I'm just a user of numerical models. So I don't believe I'm qualified/able to comment two some things. Specifically: on Functionality-Performance - the article claims that the performance is better than an Excel workbook - but I did not test this claim and don't think I could do so in the time I can devote to this review. on Documentation -Functionality documentation - I don't know what API method documentation is and cannot comment.

Thanks for providing your review along with caveats @amandersillinois. To weigh in a bit, API documentation is basically the explanations associated with individual functions, describing what the function does, what parameters it takes as inputs, and what it returns. [This](https://numpy.org/doc/1.26/reference/generated/numpy.mean.html#numpy.mean) is the API documentation for `numpy.mean` for example.

@mgcooper - any chance you've got an update for us?
