### 2024-02-19 #6191 @mgcooper
[10.21105.joss.06191.pdf](https://github.com/openjournals/joss-reviews/files/14336903/10.21105.joss.06191.pdf)
It was really simple to install pywaterflood and run the examples. I did a bare bones global install using pip and a "global" install using conda, both went smoothly. I then ran the "simple example" directly in a python repl (please see comments below on the simple example), and I ran the .ipynb notebook in an interactive jupyter session in vscode. All went smoothly, but I was surprised at how slow the code is. I don't know how computationally intensive the methods underlying the "simple example" are, but I would find it hard to use this software in a research setting (e.g., estimate parameter uncertainty) if it takes several minutes to run one simple example. Curious if the speed is typical for the field. For reference, the "7-minutes to pywaterflood" ran in about 8 min on my 2019 Quad-Core Macbook Pro with 16 GB ram. No worries if this is simply what it is, just curious if attention was paid to performance.

The "simple example" is also a bit underwhelming (although the notebook is much better). It produces no graphics or table to visualize the results, print anything to the screen, or provide any feedback to users. The main Readme say "Finally, run CRM and check the predictions and residuals" - but where is the "check" done? Clearly a user can do what I did - import pyplot and plot q_hat etc., but I had no idea what the data was or how to label the plot etc. All to say - it would be simple to port a plot command used in the notebook or print something to the screen to provide users with feedback in the repl or enable copying the simple example to a script.

I agree with @amandersillinois that the writing is a bit terse, even for JOSS. Please see the general remarks below and the comments in the attached paper.

General remarks following the JOSS review criteria:

- Please define all jargon in the Summary and provide the reader with a broader understanding of the topic. See the JOSS paper guidelines: “summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.”
- Please consider adding a reference to at least one paper following each sentence in the Summary.
- Given the lack of software that specifically performs well connectivity (see comment below in "minor suggestions"), would it be helpful to discuss software that overlaps with but stops short of the functionality offered by pywaterflood?
- Please address this: "Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it.". This and the suggestions below address the "Substantial scholarly effort" criteria.
- Please double check that there are no other public packages or software for waterfloods and/or CRM. Even if not python, I believe they still should be mentioned. This will address "description of how this software compares to other commonly-used packages in this research area."

Minor suggestions:

- Please add a note to the main readme explaining how to run the tests. I recall this is strongly suggested by JOSS at the submission stage, and seems it should be included for anyone who wishes to contribute to the software or develop on top of it.
- I did a very quick search for "python well connectivity analysis" and was unable to find any other publicly available python packages for CRM. I did find:
	- https://github.com/ICWallis/T21-Tutorial-WellTestAnalysis
	- https://github.com/Skoltech-CHR/DeepField?tab=readme-ov-file
Neither appear directly applicable, just mentioning in case the researchers have developed anything else, or if DeepField or a  similar package is worth mentioning (see general comment). No worries if these are totally unrelated, mainly want to ensure nothing was overlooked.
- I am not sure if the python notebook is within the scope of this review, but I suggest not describing a method as having been "perfected".
