### 2024-02-21 frank1010111/pywaterflood#80 @frank1010111
JOSS review: edit docs to follow suggestions
1. The "simple example" is also a bit underwhelming (although the notebook is much better). It produces no graphics or table to visualize the results, print anything to the screen, or provide any feedback to users.
2. The writing is a bit terse, even for JOSS. Please see the general remarks below and the comments in the attached paper.
3. Please add a note to the main readme explaining how to run the tests. I recall this is strongly suggested by JOSS at the submission stage, and seems it should be included for anyone who wishes to contribute to the software or develop on top of it.
4. I suggest not describing a method as having been "perfected".

_Shortened from review originally posted by @mgcooper in https://github.com/openjournals/joss-reviews/issues/6191#issuecomment-1953164664_
