### 2024-02-21 #6191 @frank1010111
Thank you for your review and suggestions.

You can track my changes to the documentation in the above-linked issue.

Two points that do not have to go into the documentation:

1. As far as the speed of the solution, the [fitting is fully parallelized](https://pywaterflood.readthedocs.io/en/latest/autoapi/pywaterflood/crm/index.html#pywaterflood.crm.CRM.fit). If you use the num_cores parameter, you can get significant speedup for this procedure. I will note that in the examples. The computationally-intensive parts are implemented in Rust.
2. The only other python package for CRM I am aware of is https://github.com/leleony/proxy-crm, which appears to have copied significant portions of pywaterflood, with some attribution.
