### 2024-02-21 36748db
:memo: num_cores param for docstring for CRM.fit
