### 2024-02-21 6a52e1f
:memo: expand simple example, include nox in README
