### 2024-02-22 4125d12
:memo: JOSS paper edits
