+++ paper/paper.md
@@ -21,9 +21,11 @@ bibliography: paper.bib
 
 # Summary
 
-Well connectivity analysis has many applications for subsurface energy, from waterfloods to CO$_2$ floods to geothermal energy projects. After completing a well connectivity analysis, reservoir managers can better allocate injections and plan well interventions. Capacitance Resistance Models are useful for performing well connectivity analysis with limited information about the geology of the reservoirs involved. They are so-called because the equations describing well influence mimic a network of capacitors and resistors.
+Well connectivity analysis has many applications for subsurface energy, covering any project where nearby wells are expected to influence one another, whether they are in oil or gas fields, geothermal fields, or an aquifer. After completing a well connectivity analysis, reservoir managers know the degree and time-dependence of influence of one well's behavior on another's. With this knowledge, they can better allocate injections and plan well interventions. Capacitance Resistance Models are useful for performing well connectivity analysis with limited information about the geology of the reservoirs involved [@yousef2006capacitance]. They are so-called because the equations describing well influence mimic a network of capacitors and resistors [@holanda2018].
 
-`Pywaterflood` is a Python package that uses Capacitance Resistance Modeling (CRM) to estimate well connectivity. The `CRM` submodule forms the bulk of this package. It can perform CRM with differing levels of complexity, from assuming that producing and injecting wells share one universal time constant, to each producer has the same time constant with all injectors, to each producer-injector pair has an its own time constant. CRM was developed by @yousef2006capacitance. The `MPI` submodule uses a geometrical model of well influence [@valko2000development], extended and applied to reservoirs with both injecting and producing wells [@kaviani2010inferring].
+`Pywaterflood` is a Python package that uses Capacitance Resistance Modeling (CRM) to estimate well connectivity. It is a portmanteau of "Python" and "waterflood," where a waterflood is an oil reservoir with water injection designed to increase reservoir pressure and move oil towards producing wells. The `CRM` submodule forms the bulk of this package. It can perform CRM with differing levels of complexity, from assuming that producing and injecting wells share one universal time constant, to each producer has the same time constant with all injectors, to each producer-injector pair has an its own time constant. CRM was developed by @yousef2006capacitance.
+
+The `MPI` submodule uses a geometrical model of well influence [@valko2000development], extended and applied to reservoirs with both injecting and producing wells [@kaviani2010inferring]. As a geometrical model, it can assist in planning reservoirs before any production or injection has begun.
 
 # Statement of need
 
@@ -34,19 +36,33 @@ Interwell connectivity analysis is important for understanding the geology of su
 3. Measuring the per-well effectiveness of fluid injection
 4. Optimizing injection, either through redirecting fluid to different wells or to inform the placement of new wells
 
-`Pywaterflood` uses a reduced-physics model to match connections between injecting and producing wells. As explained in @holanda2018, capacitance-resistance modeling provides a method for connectivity analysis more sophisticated than empirical decline analysis, but more approachable than full reservoir simulation.
+`Pywaterflood` uses a reduced-physics model to match connections between injecting and producing wells. As explained in @holanda2018, CRM provides a method for connectivity analysis that is more sophisticated than empirical decline analysis but also more approachable than full reservoir simulation.
 
[…]
