### 2024-03-01 #6191 @elbeejay
@frank1010111 it sounds like you are waiting on @mgcooper to merge https://github.com/frank1010111/pywaterflood/pull/81 and then close https://github.com/frank1010111/pywaterflood/issues/80. Is that right?
