### 2024-03-01 bae6513
:memo: fix typos in docstrings
