### 2024-03-05 642e9db
:bug: fix CrmCompensated predict function
