### 2024-03-05 dcbb788
:test_tube: add test that would have tripped on CrmCompensated.predict
