+++ .github/workflows/tests.yml
@@ -40,6 +40,7 @@ jobs:
         run: pipx run nox -s tests
       - name: "Upload coverage to Codecov"
         uses: codecov/codecov-action@v4
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         with:
           fail_ci_if_error: true
-          token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
