### 2024-03-08 #6191 @elbeejay
@frank1010111 I'd ask that you consider including the full name for the MPI acronym (I realize you are introducing the module), but still think it could be helpful for the casual reader to see the phrase "multiwell productivity index", and it might help folks find the paper when conducting a search using that language.

I've checked too and it doesn't seem like those two references that the bot flagged have DOIs, so we will do without.

At this time I will generate a post-review checklist below that has a few final steps for you to complete/confirm for me. Once we've finished these steps I'll be able to recommend this paper for publication and an editor-in-chief will give it one final look before publishing it.
