+++ .github/workflows/tests.yml
@@ -34,7 +34,7 @@ jobs:
         python-version: ["3.8", "3.12"]
         include:
           - python-version: pypy-3.10
-            runs-on: ubuntu-latest
+            os: Ubuntu
       fail-fast: false
     steps:
       - uses: actions/checkout@v4
