### 2024-03-09 25a6518
:hammer: bump version to 0.3.3
remove python 3.7 support
