+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pywaterflood"
-version = "0.3.2"
+version = "0.3.3"
 description = "Physics-inspired waterflood performance modeling"
 authors = [{ name = "Frank Male", email = "frmale@utexas.edu" }]
 license = { file = "LICENSE" }
@@ -14,7 +14,7 @@ classifiers = [
     "Topic :: Scientific/Engineering",
     "Topic :: Scientific/Engineering :: Hydrology",
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 dependencies = [
     "numpy >=1.21",
     "scipy >=1.4",
