### 2024-03-09 7be808c
:hammer: update cargo version to 0.3.3
