### 2024-03-09 b8a2c17
:test_tube: fix match statement for test_predict_fails
