### 2024-03-11 b0fd798
:memo: fix docstring example in CrmCompensated
