### 2024-03-11 9247a1e
:test_tube: fixture and other fun stuff for testing CrmCompensated
