+++ .github/workflows/publish-to-pypi.yml
@@ -104,6 +104,6 @@ jobs:
         uses: pypa/gh-action-pypi-publish@release/v1
 
       - name: Upload artifacts to release
-        uses: softprops/action-gh-release@v1
+        uses: softprops/action-gh-release@v2
         with:
           files: dist/*
