+++ .github/workflows/publish-to-pypi.yml
@@ -18,13 +18,12 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-22.04, windows-2022, macos-12]
+        os: [ubuntu-22.04, windows-2022, macos-12, macos-14]
     env:
       # rust is not available for musl linux on i686. see https://github.com/rust-lang/rustup/issues/2984
       CIBW_SKIP: "*-musllinux_i686"
       CIBW_BEFORE_ALL_LINUX: "curl https://sh.rustup.rs -sSf | sh -s -- --default-toolchain stable -y"
       CIBW_ENVIRONMENT: 'PATH="$PATH:$HOME/.cargo/bin"'
-      CIBW_ARCHS_MACOS: "x86_64 arm64"
 
     steps:
       - uses: actions/checkout@v4
