### 2024-03-14 #6191 @frank1010111
I've released the latest version of pywaterflood as v0.3.3. Here is the DOI: https://zenodo.org/doi/10.5281/zenodo.6569905
