### 2024-03-15 #6191 @frank1010111
Hi @kthyng, I've updated the archive to match title and author (and ORCID) to the JOSS paper.
