### 2024-03-19 4e9d0a8
:memo: clarify paper grammar following JOSS editor comments
